% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKNMI.R
\name{get_knmi_daydata}
\alias{get_knmi_daydata}
\title{get_knmi_daydata}
\usage{
get_knmi_daydata(start, end, vars = "Q:TG", stns = "310", prefer_json = TRUE)
}
\arguments{
\item{start}{start date (yyyymmdd)}

\item{end}{end date (yyyymmdd)}

\item{vars}{variables (see: https://www.knmi.nl/kennis-en-datacentrum/achtergrond/data-ophalen-vanuit-een-script)}

\item{stns}{stations (see: https://www.knmi.nl/kennis-en-datacentrum/achtergrond/data-ophalen-vanuit-een-script)}

\item{prefer_json}{toggles csv or json output}
}
\value{
KNMI daily average values in memory
}
\description{
get_knmi_daydata
}
\examples{
# example
knmi_daily <- get_knmi_daydata(start = "19950101", end = "20250101", vars = c("FG:FHX:FHXH"), prefer_json = FALSE)

}
