% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_psmsl_data.R
\name{get_psmsl_station_table}
\alias{get_psmsl_station_table}
\title{get_psmsl_station_table}
\usage{
get_psmsl_station_table(url = "https://psmsl.org/data/obtaining/index.php")
}
\arguments{
\item{url}{url of psmsl station table page. Default is https://psmsl.org/data/obtaining/index.php}
}
\value{
dataframe with station information including lat and lon. The information is scraped from the web page.
}
\description{
get_psmsl_station_table
}
\examples{
stationinfo <- get_psmsl_station_table()
}
