% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtsm_nc.R
\name{read_gtsm_nc}
\alias{read_gtsm_nc}
\title{read_gtsm_nc}
\usage{
read_gtsm_nc(nc_path, stations_selected)
}
\arguments{
\item{nc_path}{path to netcdf output generated by the Global Tide and Surge Model (GTSM).}

\item{stations_selected}{GTSM ID's of monitoring stations in output file.}
}
\value{
GTSM netcdf file in memory
}
\description{
read_gtsm_nc
}
