% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKNMI.R
\name{read_knmi_hourly}
\alias{read_knmi_hourly}
\title{read_knmi_hourly
Reads files as downloaded via https://www.knmi.nl/nederland-nu/klimatologie/uurgegevens}
\usage{
read_knmi_hourly(filename, n_max = Inf)
}
\arguments{
\item{filename}{file name of downloaded file}

\item{n_max}{number of lines to read}
}
\value{
KNMI hourly average values in memory
}
\description{
read_knmi_hourly
Reads files as downloaded via https://www.knmi.nl/nederland-nu/klimatologie/uurgegevens
}
\examples{
# example
f <- system.file("extdata", "dekooy_uurgeg_wind_test.txt", package = "rsealevel")
read_knmi_hourly(f)
}
