% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_psmsl_data.R
\name{read_monthly_psmsl_csv}
\alias{read_monthly_psmsl_csv}
\title{read_monthly_psmsl_csv}
\usage{
read_monthly_psmsl_csv(station_nr)
}
\arguments{
\item{station_nr}{psmsl station number}
}
\value{
dataframe containing sea level data for the requested station
}
\description{
read_monthly_psmsl_csv
}
\examples{
read_monthly_psmsl_csv(c(20, 22, 23, 24, 25, 32))
}
