% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tidal_components_csv.R
\name{read_tidal_components_csv}
\alias{read_tidal_components_csv}
\title{read_tidal_components_csv}
\usage{
read_tidal_components_csv(
  files = NA,
  filesdir =
    "p:/11202493--systeemrap-grevelingen/1_data/Wadden/ddl/calculated/TA_filtersurge"
)
}
\arguments{
\item{files}{Specify files to read. Alternatively specify a directory to read all files from.}

\item{filesdir}{Specify directory to read all files from. Either files or filesdir needs to be specified.}
}
\value{
dataframe with amplitudes and phases for all stations and tidal components in the files read.
}
\description{
read_tidal_components_csv
}
