% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMainStationInfo.R
\name{readMainStationInfo}
\alias{readMainStationInfo}
\title{readMainStationInfo}
\usage{
readMainStationInfo(filepath = "")
}
\arguments{
\item{filepath}{path to main_stations.json file}
}
\value{
dataframe with station info of 6 main stations as used in the Dutch Sea Level Monitor
}
\description{
readMainStationInfo
}
\examples{
readMainStationInfo()
}
