% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMainStationInfo.R
\name{readMainStationLocations}
\alias{readMainStationLocations}
\title{Title}
\usage{
readMainStationLocations(path = "")
}
\arguments{
\item{path}{path to file NLstations.csv}
}
\value{
dataframe with locations for Dutch main tide gauges only. The information has been retrieved from PSMSL.
}
\description{
Title
}
